move fs-related functions
